+++ security/oc_doxm.c
@@ -77,7 +77,7 @@ oc_set_doxm(oc_doxm_method_t oxmsel)
 }
 
 void
-oc_sec_doxm(int device, oc_doxm_method_t oxmsel)
+oc_sec_doxm(size_t device, oc_doxm_method_t oxmsel)
 {
   doxm[device].oxmsel = oxmsel;
   doxm[device].sct = 1;
@@ -88,7 +88,7 @@ oc_sec_doxm(int device, oc_doxm_method_t oxmsel)
 }
 
 void
-oc_sec_encode_doxm(int device)
+oc_sec_encode_doxm(size_t device)
 {
   int oxms[1] = { doxm[device].oxmsel };
   char uuid[OC_UUID_LEN];
@@ -109,7 +109,7 @@ oc_sec_encode_doxm(int device)
 }
 
 oc_sec_doxm_t *
-oc_sec_get_doxm(int device)
+oc_sec_get_doxm(size_t device)
 {
   return &doxm[device];
 }
@@ -122,7 +122,7 @@ get_doxm(oc_request_t *request, oc_interface_mask_t interface, void *data)
   case OC_IF_BASELINE: {
     char *q;
     int ql = oc_get_query_value(request, "owned", &q);
-    int device = request->resource->device;
+    size_t device = request->resource->device;
     if (ql > 0 &&
         ((doxm[device].owned == 1 && strncasecmp(q, "false", 5) == 0) ||
          (doxm[device].owned == 0 && strncasecmp(q, "true", 4) == 0))) {
@@ -143,11 +143,11 @@ get_doxm(oc_request_t *request, oc_interface_mask_t interface, void *data)
 }
 
 bool
-oc_sec_decode_doxm(oc_rep_t *rep, bool from_storage, int device)
+oc_sec_decode_doxm(oc_rep_t *rep, bool from_storage, size_t device)
 {
   oc_sec_pstat_t *ps = oc_sec_get_pstat(device);
   oc_rep_t *t = rep;
-  int len = 0;
+  size_t len = 0;
 
   while (t != NULL) {
     len = oc_string_len(t->name);
